refactor: move IPC subscription logic into ipc module
Extract the socket listener subscription from app.rs into
ipc::subscription(), keeping app.rs clean and the IPC logic
self-contained.
